+++ typings/custom-typings.d.ts
@@ -12,3 +12,7 @@ declare module '*.svg' {
     const content: any;
     export default content;
 }
+
+declare interface Window {
+    DocumentTouch: any;
+}